Refactor the SmartSnapshot Coordinator for Postgres, abstract the common piece of reading the snapshot info, failure handling etc to the core
